requirements.txt
# --------------------
# CORE FRAMEWORK
# --------------------
fastapi==0.119.0
uvicorn==0.37.0
starlette==0.48.0

# --------------------
# DATABASE & ORM
# --------------------
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
greenlet==3.2.4

# --------------------
# SCHEMAS & VALIDATION (Compatible versions!)
# --------------------
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.6.1

# --------------------
# EMAILS & HTML TEMPLATES
# --------------------
fastapi-mail==1.5.7
email-validator==2.3.0
jinja2==3.1.6
python-multipart==0.0.20
aiosmtplib==4.0.2
sendgrid==6.11.0

# --------------------
# AUTHENTICATION & SECURITY
# --------------------
bcrypt==5.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cryptography==46.0.3
python-jose==3.5.0
passlib==1.7.4

# --------------------
# ENVIRONMENT & UTILITIES
# --------------------
python-dotenv==1.0.1
typing-extensions==4.15.0
annotated-types==0.7.0

# --------------------
# NETWORKING / ASYNC SUPPORT
# --------------------
anyio==4.11.0
httpx==0.27.2
sniffio==1.3.1
websockets==15.0.1

# --------------------
# LOGGING / MONITORING
# --------------------
sentry-sdk==2.41.0
colorama==0.4.6
rich==14.2.0
watchfiles==1.1.0
python-json-logger==3.3.0

# --------------------
# DEV & TEST UTILITIES
# --------------------
pytest==8.4.2
ruff==0.13.2
click==8.3.0

# --------------------
# OPTIONAL TOOLING
# --------------------
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
Authlib==1.6.5
supabase==2.10.0
requests==2.31.0
apscheduler==3.10.4

# --------------------
# PACKAGE MANAGEMENT UTILITIES
# --------------------
virtualenv==20.34.0
distlib==0.4.0
platformdirs==4.4.0

# --------------------
# ML Dependencies
# --------------------
# scikit-learn==1.3.2
# numpy==1.24.3


scikit-learn>=1.5.0,<1.6.0
numpy>=1.23.0
scipy>=1.8.0
openai==1.54.3
sib-api-v3-sdk==7.6.0
